Move geometry dataclasses into models_geometry.py
Split the geometry definitions (PixelClass, Image, SubVolumeType,
GeometryClass, SubVolume, SurfaceClass, Geometry) out of models.py into
models_geometry.py, alongside models_base.py and models_math.py. This lets
geometry- and math-only third-party packages depend on those datamodels
without pulling in biomodel constructs.

Application-level structure mapping (StructureMapping / CompartmentMapping)
stays with the biomodel in models.py, as it relates the model to a
geometry. models.py imports Geometry/GeometryClass for its own fields; the
reader/writer/__init__ import geometry types from models_geometry.

No behavior change. make check passes; tests green.
