+++ pyvcell/vcml/__init__.py
@@ -5,24 +5,26 @@ from pyvcell.vcml.models import (
     Biomodel,
     BoundaryType,
     Compartment,
-    Geometry,
-    Image,
     Kinetics,
     KineticsParameter,
     Model,
     ModelParameter,
-    PixelClass,
     Reaction,
     Simulation,
     Species,
     SpeciesMapping,
     SpeciesReference,
     SpeciesRefType,
+    VCMLDocument,
+    Version,
+)
+from pyvcell.vcml.models_geometry import (
+    Geometry,
+    Image,
+    PixelClass,
     SubVolume,
     SubVolumeType,
     SurfaceClass,
-    VCMLDocument,
-    Version,
 )
 from pyvcell.vcml.models_math import (
     Boundaries,
